Add new images and controllers for client management
This commit adds several new image files to the project, as well as three new controller classes: ControladorAltaCliente, ControladorConsultaCliente, and ControladorVentanaPrincipal. These controllers are responsible for managing client data in the application. The ControladorAltaCliente class handles adding a new client to the system, while the ControladorConsultaCliente class allows users to view and edit existing clients. Finally, the ControladorVentanaPrincipal class is responsible for controlling the main window of the application.